refactor: change content feature folder structure
Create a folder for each feature. Put the feature script as index.ts
inside the respective feature folder. Move feature components under a
new components folder in the respective feature folder. Capitilize
component names.
