+++ src/content/features/add-player-map-stats/components/StatsTable.tsx
@@ -1,10 +1,10 @@
 /** @jsx h */
 import { h } from "dom-chef";
 import { ReactNode } from "react";
-import { ACTIVE_MAP_POOL, EMPTY_STATS, ESCL } from "../../shared/consts";
-import { colors } from "../../shared/theme";
-import { MapStats } from "../../shared/types/stats";
-import createMapStatsCell from "./mapStatsCell";
+import { ACTIVE_MAP_POOL, EMPTY_STATS, ESCL } from "../../../../shared/consts";
+import { colors } from "../../../../shared/theme";
+import { MapStats } from "../../../../shared/types/stats";
+import createMapStatsCell from "./StatsCell";
 
 type _h = typeof h; // needed to prevent TSeslint from removing h import
